Added CameraUtils dependency
- Removed Camera modifications logic
- Migrated LayerUtils.NoteLayer to CameraUtils.VisibilityLayer enum
- Renamed GameCameraManager to WatermarkManager
