+++ CustomNotes/CustomNotes.csproj
@@ -45,6 +45,9 @@
       <HintPath>$(BeatSaberDir)\Plugins\BSML.dll</HintPath>
       <Private>False</Private>
     </Reference>
+    <Reference Include="CameraUtils">
+      <HintPath>$(BeatSaberDir)\Plugins\CameraUtils.dll</HintPath>
+    </Reference>
     <Reference Include="Colors, Version=0.0.0.0, Culture=neutral, processorArchitecture=MSIL">
       <Private>False</Private>
       <HintPath>$(BeatSaberDir)\Beat Saber_Data\Managed\Colors.dll</HintPath>
@@ -139,7 +142,7 @@
     <Compile Include="Managers\CustomBurstSliderController.cs" />
     <Compile Include="Managers\CustomNoteController.cs" />
     <Compile Include="Managers\CustomNotesViewManager.cs" />
-    <Compile Include="Managers\GameCameraManager.cs" />
+    <Compile Include="Managers\WatermarkManager.cs" />
     <Compile Include="Managers\MenuButtonManager.cs" />
     <Compile Include="Managers\CustomNoteManager.cs" />
     <Compile Include="Overrides\CustomNoteColorNoteVisuals.cs" />
